requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
distro==1.9.0
Django==4.2.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
executing==2.0.0
fastapi==0.115.6
fastjsonschema==2.18.1
filelock==3.12.3
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.41
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.26.2
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
gTTS==2.3.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.4
importlib-metadata==7.0.1
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyter_server==2.8.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
mysql-connector-python==8.0.33
mysqlclient==2.2.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.4.2
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.3
openai==1.7.2
opencv-python==4.7.0.72
opencv-python-headless==4.10.0.84
openpyxl==3.1.2
overrides==7.4.0
packaging==23.2
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
phonenumbers==8.13.19
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
playsound==1.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.6
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.19
pytz==2023.3
pywin32==306
pywinpty==2.0.12
PyYAML==6.0
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.10.6
rsa==4.9
scikit-learn==1.3.1
scipy==1.10.1
seaborn==0.13.0
Send2Trash==1.8.2
six==1.16.0
sketchpy==0.1.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SpeechRecognition==3.10.0
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.41.3
streamlit==1.30.0
svg.path==6.2
svgpathtools==1.6.0
svgwrite==1.4.3
sympy==1.13.1
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
torch==2.5.1
torchvision==0.20.1
tornado==6.3.3
tqdm==4.65.0
traitlets==5.11.2
turtle==0.0.1
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
ultralytics==8.3.47
ultralytics-thop==2.0.12
uri-template==1.3.0
urllib3==1.26.15
uvicorn==0.32.1
validators==0.22.0
virtualenv==20.24.4
watchdog==3.0.0
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
wordcloud==1.9.2
yarl==1.9.1
zipp==3.17.0
